BatchedDot: move to theano/tensor/blas.py
